requirements-gpu.txt
pynput
opencv-python
pywin32
pillow
numpy
paddlepaddle-gpu
paddleocr
setuptools
pyyaml